Make data unique for each user

+++ controllers/categoryController.js
@@ -6,7 +6,6 @@ const asyncHandler = require("express-async-handler");
 let categoryData = []
 let itemData = []
 
-
 exports.add_category = asyncHandler(async (req, res, next) => {
     const name = req.body.categoryName;
     const description = req.body.categoryDescription;
@@ -15,16 +14,17 @@ exports.add_category = asyncHandler(async (req, res, next) => {
         name: name,
         description: description,
         URL: URL,
+        owner: req.user._id,
         expiresAfter: new Date()
     });
     res.redirect("/categories");
 });
 
 exports.render_categories = asyncHandler(async (req, res, next) => {
-    categoryData = await Category.find({}).exec();
-    itemData = await Item.find({}).exec();
+    categoryData = await Category.find({ owner: req.user._id }).exec();
+    itemData = await Item.find({ owner: req.user._id }).exec();
     const page = "categories"
-    res.render('layout', { title: 'Categories', categories: categoryData, items: itemData, page: page});
+    res.render('layout', { title: 'Categories', categories: categoryData, items: itemData, page: page });
 });
 
 exports.update_category = asyncHandler(async (req, res, next) => {
@@ -32,11 +32,12 @@ exports.update_category = asyncHandler(async (req, res, next) => {
     const newName = req.body.categoryName;
     const newDescription = req.body.categoryDescription;
     const newURL = req.body.categoryURL;
-    await Category.findOneAndUpdate({ _id: categoryID, }, {
+    await Category.findOneAndUpdate({ owner: req.user._id, _id: categoryID, }, {
         $set: {
             name: newName,
             description: newDescription,
             URL: newURL,
+            owner: req.user._id,
             expiresAfter: new Date()
         }
     });
@@ -45,7 +46,7 @@ exports.update_category = asyncHandler(async (req, res, next) => {
 
 exports.delete_category = asyncHandler(async (req, res, next) => {
     const categoryID = req.body.categoryID;
-    await Category.deleteOne({ _id: categoryID });
+    await Category.deleteOne({ owner: req.user._id, _id: categoryID });
     res.redirect("/categories");
 });
 
@@ -58,14 +59,14 @@ exports.render_add_form = asyncHandler(async (req, res, next) => {
 exports.render_edit_form = asyncHandler(async (req, res, next) => {
     const categoryID = req.body.categoryID;
     const page = "edit-category"
-    const category = await Category.findById(categoryID).exec();
+    const category = await Category.findOne({ owner: req.user._id, _id: categoryID }).exec();
     res.render('layout', { title: 'Edit Category', category: category, page: page, categoryID: categoryID });
 });
 
 exports.render_items = asyncHandler(async (req, res, next) => {
-    itemData = await Item.find({}).exec();
+    itemData = await Item.find({ owner: req.user._id }).exec();
     const categoryID = req.params.objectId;
-    const currentCategory = await Category.findById(categoryID).exec();
+    const currentCategory = await Category.findOne({ owner: req.user._id, _id: categoryID }).exec();
     const page = "items"
     res.render('layout', { title: 'Items', currentCategory: currentCategory, items: itemData, page: page });
 });
